create a review for a spot based on the spot's id started

+++ backend/db/migrations/20230217223608-create-reviews.js
@@ -27,7 +27,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       stars: {
-        type: Sequelize.DECIMAL,
+        type: Sequelize.INTEGER,
         allowNull: false,
       },
       createdAt: {
